requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==18.10.1
h3==3.7.4
psycopg2-binary==2.9.3
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
